oriscv: read instruction addresses, too
For a more precise and insightful analysis of register chains, it will
be useful to pinpoint the address of those chains.  This requires a
change in the input format.  It used to be the raw hex value of
instructions; this becomes the hex addr, followed by a colon, then the
hex value of the instruction.  The readme has been updated to reflect
this change.

To be really future-proof, change the way parameters are parsed (oriscv
now uses the Arg module from the stdlib), and add a (currently unused)
parameter to set the instruction window size for the register chain analysis.

Signed-off-by: Alban Gruin <alban.gruin@univ-tlse3.fr>